Refactored CMake build.
After a fresh clone of this repository, the code would not build. I
buckled up and went down the rabbit hole that is learning CMake. After a
lot (like, a ton) of Googling, I finally figured out enough to fix
things and also clean up a few things too. Notably,

- Before, compiled protos were included in the repo. This meant that
  anyone building the project needed to have the same exact `protoc`
  version or rebuild all the protos manually. Two unattractive options.
  I modified things so that CMake now builds protos for us.
- I fixed includes to follow Google's style guide [1][]. Doing so
  required a good amount of CMake wrangling.
- I fixed up miscellaneous things here and there. For example, library
  dependencies weren't being set properly and warnings like "-lprotobuf
  flag ignored" were being emitted.

[1]: https://google.github.io/styleguide/cppguide.html#Names_and_Order_of_Includes